Move DocumentAfterResolution to separate file
